Refactor QuickConnect helper functions
